Add JSON parsing and batch processing to Program.cs
A Newtonsoft.Json array parsing has been added for reading the input file content. Added call to batch processing method in Program.cs with relevant arguments, enhancing the versatility and robustness of the application. This allows batch handling of data import tasks and provides progress feedback if not running in silent mode.