Add progress update feature to SQL import process
A new optional feature has been added to the SQL import process that allows for tracking the progress of ongoing batch processes. This includes the addition of a timer and an event handler onUpdate, which gets invoked with processed and total counts periodically and at process end. Useful for long running tasks, allowing real-time tracking of process progress.