Add connection port option to CLI
A new command-line interface option has been added to specify the server connection port. The default value is set to 3306 if not explicitly provided. A helper function `GetConnectionPort` is also added to parse the port value and check its validity, throwing an error if the inserted port value is not an integer.