Add test mode to bulk import process
Test mode was added to the bulk import process in the BulkImportSQL program. This mode will not execute data insertions but instead generate and pring SQL insert queries. Ten lines are displayed on the screen and the remaining lines get written into an SQL file for manual reviews or runs.